[cloud](storage vault) Rewrite vaults regression test and fix two error (#42411)
* Rewrite vaults regression test
* Fix show storage vault stmt `default` column incorrectly
* Fix create table cannot find storage vault after creating the vault
